+++ app/javascript/components/TaskShow.tsx
@@ -1,0 +1,19 @@
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Link,
+  useParams
+} from "react-router-dom";
+
+const TaskShow = () => {
+  let { id } = useParams();
+
+  return (
+    <div>
+      <p> {id}</p>
+    </div>
+  );
+}
+
+export default TaskShow;
