+++ package.json
@@ -9,6 +9,7 @@
     "postcss": "^8.4.16",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "tailwindcss": "^3.1.8"
   },
   "scripts": {
